requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
azure-identity==1.23.0
pytest==8.4.1
